Refactored Google Calendar authentication out of gcal/index.js. Prototyped front-end API.
