+++ .gitignore
@@ -1,2 +1,3 @@
 node_modules
 client_secret.json
+npm-debug.log
